save system overhaul (w items)
sets the foundation for the data saving system. data is saved in xml trees, and each scene has its own xml file to hold its data. data is read in at a scene's load and written back to the file upon exiting the scene. functionality for items on the ground is included